update HttpAcceptorRule to be able to set any ExpiringState instance
 - update multiple auth test
